+++ fonduer/candidates/models/__init__.py
@@ -1,3 +1,15 @@
-from fonduer.candidates.models.candidate import Candidate, Marginal, candidate_subclass
+from fonduer.candidates.models.candidate import Candidate, candidate_subclass
+from fonduer.candidates.models.image import Image, TemporaryImage
+from fonduer.candidates.models.implicitspan import ImplicitSpan, TemporaryImplicitSpan
+from fonduer.candidates.models.span import Span, TemporarySpan
 
-__all__ = ["Candidate", "Marginal", "candidate_subclass"]
+__all__ = [
+    "Candidate",
+    "Image",
+    "ImplicitSpan",
+    "Span",
+    "TemporaryImage",
+    "TemporaryImplicitSpan",
+    "TemporarySpan",
+    "candidate_subclass",
+]