+++ fonduer/features/core_features.py
@@ -1,4 +1,4 @@
-from fonduer.parser.models import ImplicitSpan, TemporarySpan
+from fonduer.candidates.models import ImplicitSpan, TemporarySpan
 
 FEAT_PRE = "CORE_"
 DEF_VALUE = 1