+++ fonduer/parser/models/__init__.py
@@ -1,34 +1,17 @@
-from fonduer.parser.models.context import (
-    Cell,
-    Context,
-    Document,
-    Figure,
-    Image,
-    ImplicitSpan,
-    Phrase,
-    Span,
-    Table,
-    TemporaryImage,
-    TemporaryImplicitSpan,
-    TemporarySpan,
-    Webpage,
-    construct_stable_id,
-    split_stable_id,
-)
+from fonduer.parser.models.context import Context, construct_stable_id, split_stable_id
+from fonduer.parser.models.document import Document
+from fonduer.parser.models.figure import Figure
+from fonduer.parser.models.phrase import Phrase
+from fonduer.parser.models.table import Cell, Table
+from fonduer.parser.models.webpage import Webpage
 
 __all__ = [
     "Cell",
     "Context",
     "Document",
     "Figure",
-    "Image",
-    "ImplicitSpan",
     "Phrase",
-    "Span",
     "Table",
-    "TemporaryImage",
-    "TemporaryImplicitSpan",
-    "TemporarySpan",
     "Webpage",
     "construct_stable_id",
     "split_stable_id",