+++ CHANGELOG.rst
@@ -9,6 +9,8 @@ Version 0.2.0 (coming soon...)
 * `@lukehsiao`_: Remove the futures imports, truly making Fonduer Python 3
   only. Also reorganize the codebase into submodules for each pipeline phase.
   (`#59 <https://github.com/HazyResearch/fonduer/pull/59>`_)
+* `@lukehsiao`_: Split models into individual files.
+  (`#60 <https://github.com/HazyResearch/fonduer/pull/60>`_)
 
 Version 0.1.8
 -------------